Fixed integer overflow with HTTP cache - tradeid
-Fixed integer overflow with HTTP cache - tradeid (Okcoin BTC/CNY's
tradeid have exceed 2.147 billion as of 12/03/2015

+++ BitBot/src/bitbot/cache/tickers/HTTP/TickerHistory_CampBX.java
@@ -31,7 +31,7 @@ public class TickerHistory_CampBX implements TickerHistoryInterface {
     }
 
     @Override
-    public TickerHistoryData connectAndParseHistoryResult(TickerCacheTask.TickerCacheTask_ExchangeHistory _TickerCacheTaskSource, String ExchangeCurrencyPair, String ExchangeSite, String CurrencyPair, long LastPurchaseTime, int LastTradeId) {
+    public TickerHistoryData connectAndParseHistoryResult(TickerCacheTask.TickerCacheTask_ExchangeHistory _TickerCacheTaskSource, String ExchangeCurrencyPair, String ExchangeSite, String CurrencyPair, long LastPurchaseTime, long LastTradeId) {
         String Uri = "http://CampBX.com/api/xticker.php";
         String Result = HttpClient.httpGet(Uri, "");
