+++ BitBot/src/bitbot/cache/tickers/HTTP/TickerHistory_OkcoinInternational.java
@@ -30,7 +30,7 @@ public class TickerHistory_OkcoinInternational implements TickerHistoryInterface
     }
 
     @Override
-    public TickerHistoryData connectAndParseHistoryResult(TickerCacheTask.TickerCacheTask_ExchangeHistory _TickerCacheTaskSource, String ExchangeCurrencyPair, String ExchangeSite, String CurrencyPair, long LastPurchaseTime, int LastTradeId) {
+    public TickerHistoryData connectAndParseHistoryResult(TickerCacheTask.TickerCacheTask_ExchangeHistory _TickerCacheTaskSource, String ExchangeCurrencyPair, String ExchangeSite, String CurrencyPair, long LastPurchaseTime, long LastTradeId) {
         String Uri;
 
         if (CurrencyPair.contains("Futures")) {
@@ -76,7 +76,7 @@ public class TickerHistory_OkcoinInternational implements TickerHistoryInterface
                     float price = Float.parseFloat(obj.get("price").toString());
                     long date = (long) Double.parseDouble(obj.get("date").toString()) * 1000l;
                     final TradeHistoryBuySellEnum type = obj.get("type").toString().equals("buy") ? TradeHistoryBuySellEnum.Buy : TradeHistoryBuySellEnum.Sell; // buy sell
-                    int tid = Integer.parseInt(obj.get("tid").toString());
+                    long tid = Long.parseLong(obj.get("tid").toString());
 
                     // Initialize last purchase time if neccessary
                     if (LastPurchaseTime == 0) {
